# FE/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sports Center Management API",
    "version": "1.0.0",
    "description": "Backend API for Sports Center Management System – Flow 1 (User & Membership), Flow 2 (Class & Schedule), Flow 3 (Payment & Report)"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://sports-center-management-system.onrender.com/api/v1",
      "description": "Production Server (Render)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request (validation or malformed body)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Validation failed",
                "errors": [
                  {
                    "field": "email",
                    "message": "Invalid email address"
                  }
                ]
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, invalid or expired token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Unauthorized: invalid or expired token"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient role permissions",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Forbidden: insufficient permissions"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Record not found"
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "Duplicate value or business conflict",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Duplicate value for: email"
              }
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Internal server error"
              }
            }
          }
        }
      },
...
      "UserListOk": {
        "description": "Paginated list of users (compact example)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": true,
                "message": "Users retrieved successfully",
                "data": [
                  {
                    "id": "b311dfe5-8575-4ed4-9efd-8bc81a149f14",
                    "email": "staff@sportscenter.com",
                    "fullName": "Jane Doe",
                    "phone": "0900000002",
                    "gender": "FEMALE",
                    "dateOfBirth": null,
                    "role": "STAFF",
                    "isActive": true,
                    "createdAt": "2026-09-11T14:20:14.910Z",
                    "memberProfile": null,
                    "coachProfile": null,
                    "managerProfile": null
                  }
                ],
                "pagination": {
                  "page": 1,
                  "limit": 10,
                  "total": 8,
                  "totalPages": 1
                }
              }
            }
          }
        }
      },
      "UserCreated": {
        "description": "User created (member/coach/manager gets its profile)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": true,
                "message": "User created successfully",
                "data": {
                  "id": "b311dfe5-8575-4ed4-9efd-8bc81a149f14",
                  "email": "member1@example.com",
                  "fullName": "John Doe",
                  "phone": "0900000005",
                  "gender": "MALE",
                  "dateOfBirth": null,
                  "role": "MEMBER",
                  "isActive": true,
                  "createdAt": "2026-09-11T14:20:14.910Z",
                  "memberProfile": {
                    "id": "aecd9439-82e2-47da-90a2-2830bbe04dc4",
                    "trainingLevel": "BEGINNER"
                  },
                  "coachProfile": null,
                  "managerProfile": null
                }
              }
            }
          }
        }
      },
      "UserOk": {
        "description": "Single user",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": true,
                "message": "User retrieved successfully",
                "data": {
                  "id": "b311dfe5-8575-4ed4-9efd-8bc81a149f14",
                  "email": "member1@example.com",
                  "fullName": "John Doe",
                  "phone": "0900000005",
                  "gender": "MALE",
                  "dateOfBirth": null,
                  "role": "MEMBER",
                  "isActive": true,
                  "memberProfile": {
                    "id": "aecd9439-82e2-47da-90a2-2830bbe04dc4"
                  },
                  "coachProfile": null,
                  "managerProfile": null
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/users": {
      "get": {
        "summary": "List all users",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "MEMBER",
                "COACH",
                "STAFF",
                "MANAGER"
              ]
            }
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserListOk"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "summary": "Create staff, coach, manager or member account (MEMBER creates the member profile too)",
        "description": "Với `role = MEMBER`, backend tạo MemberProfile + **auto-provision một MembershipSubscription ACTIVE\nvới gói FREE** (`maxConcurrentClasses = 0`) trong cùng transaction. COACH/STAFF/MANAGER KHÔNG được cấp\nsubscription. Idempotent theo member: nếu member đã có subscription ACTIVE thì không tạo thêm.\n",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "fullName",
                  "role"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "fullName": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "MEMBER",
                      "COACH",
                      "STAFF",
                      "MANAGER"
                    ]
                  },
                  "fitnessGoal": {
                    "type": "string",
                    "description": "Role MEMBER only"
                  },
                  "trainingLevel": {
                    "type": "string",
                    "enum": [
                      "BEGINNER",
                      "INTERMEDIATE",
                      "ADVANCED"
                    ],
                    "description": "Role MEMBER only"
                  },
                  "trainingPreference": {
                    "type": "string",
                    "description": "Role MEMBER only"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/UserCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserOk"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "summary": "Update user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "MALE",
                      "FEMALE",
                      "OTHER"
                    ]
                  },
                  "dateOfBirth": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "MEMBER",
                      "COACH",
                      "STAFF",
                      "MANAGER"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserOk"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "summary": "Deactivate user (soft delete)",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserOk"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}